feat(product): implement product CRUD

+++ src/main/java/com/dunnas/lucas/supplier_orders_api/application/dto/ProductDTO.java
@@ -1,4 +1,6 @@
 package com.dunnas.lucas.supplier_orders_api.application.dto;
 
-public record ProductDTO(Long id, String name, String description, double price) {
+import java.math.BigDecimal;
+
+public record ProductDTO(Long id, String name, String description, BigDecimal price) {
 }
